requirements.txt
docopt==0.6.2
PyYAML==6.0.1
requests==2.31.0
packaging==23.2
junit-xml==1.9
